backend/requirements.txt
flask
flask_cors
flask_login
pymongo
python-dotenv
sentence_transformers
